Feat: native관련 로직 core와 utill로 파일 분리, event함수 업데이트
Changes to be committed:
modified:   src/components/ExerciseChoice/ExercisePanelBottom/ExerciseChoiceBottom.tsx
modified:   src/components/ExerciseRoutine/RecommendSection/RecommendSection.tsx
modified:   src/components/Survey/SurveyComplete/SurveyComplete.tsx
modified:   src/pages/ExerciseRoutine/ExerciseRoutine.tsx
new file:   src/utils/mobile/action.js
new file:   src/utils/mobile/core.js
modified:   src/utils/mobile/token.js
